rename cron scripts directory
